trace2: messages in new file
Move Trace2 messages into a new file. The Trace2 file was getting a bit
long and unwieldy, so this change is intended to separate it in a logical
way into two smaller files that are (hopefully) easier to parse and reason
about.
